requirements.txt
setuptools
wheel
miditok 
pretty_midi 
pandas 
numpy 
faiss-cpu 
sentence-transformers 
torch 
tqdm 
matplotlib 
streamlit
pygame
